Refactoring class to new file
